Introduce new hotkeys for copying 'values' to clipboards.
'x' will copy to PRIMARY X selection,
'X' will copy to CLIPBOARD ("^C -> ^V") X selection,
't' will insert the value to tmux's paste buffer.

'x' and 'X' uses xclip(1): the 'value' is sent to its stdin.
't' uses the tmux binary of course, and it just calls the set-buffer
command with the 'value' as its parameter.

Both binaries must be in PATH, and if they can not be executed, an error
message gets printed to the user.

Removal of the 'values' from the clipboards are not being dealt with!!!