Bump maximum password length, and treat passwords and passwords in files equally.
Don't use a different maximum length for a password read from a password
file.

Warn the user, if the read password is longer than the allowed, and that
it got truncated. Unfortunatelly, this can not be checked with
readpassphrase().

Zero out the 'key' variable at quit().

Regression tests:

Introduce new regression test for testing the maximum password length in
a password file.

Use the longest possible password for the test database.

Add test number tracking to regress/run_tests.sh; print which test
failed, and out of how many. Also, rephrase the "all ok" message.