Import OpenBSD's bcrypt implementation to use bcrypt_pbkdf().
Also, rename the Linux Makefile to represent the purpose of it. It is
still a GNU make file, but it only used on Linux, so... I don't know,
really...

Various assortments in the Linux Makefile, and extend it, to compile the
newly added bcrypt/ directory as a dependency to kc on Linux.

The bcrypt_pbkdf() KDF is not used yet in this commit, however it is
coming.