Remove -Wall and -pedantic compiler options from bcrypt.
This is interesting, because the CFLAGS variable doesn't get picked up
in bcrypt's Makefile.linux if it is not set in the shell beforehand.
However, right after I invoke the build with from the root with eg.:
$ CFLAGS='-O2 -mtune=i686' make -f Makefile.linux
... bcrypt's Makefile.linux will pick up every CFLAGS that was defined
in the root Makefile.linux, along with the additionally defined CFLAGS
variable's content.
So use this GNU make magic, to remove strings from variables. Nice.