Use options/parameters and getopt(3) with some commands.
passwd, import and xport commands now have single character options to
specify their parameters. The options are parsed with getopt(3). This
simplifies using eg. the 'xport' command, because the keychain can now
be specified without also having to specify the kdf and cipher_mode.

New function larg(), to create a local argv/argc pair to use them with
getopt in the commands' functions.

Of course documentation, manual and Changelog was adjusted accordingly.