Change/rework the exiting/quitting logic.
If the database was dirty and the user answered 'yes' to the saving
question during exiting, then check for the 'dirty' flag after calling
cmd_write(), and if kc should have saved the database but could not
(let's hope this is always signified by the dirty flag staying true :)
), then tell the user about this, and fall back to the CLI.

Furthermore, only accept the (now just) two valid answers, yes and no.
Anything else will put the user back to the CLI (instead of looping the
question infinitely until it gets a valid answer, like before).