Check for truncations with strl{cpy,cat} ...
In cmd_import.c and kc.c don't use an intermediate buffer (*buf) for
reading in the iv/salt, just read into the db_param[_new] structure
directly.

In kc.c use variables for default db parameters (kdf, cipher, mode), and
do not use the strings directly for copying or computations.

In cmd_getnum.c mark the strl{cpy,cat} functions with (void) cast, to
indicate that we don't care (yet?) that a random string could not be
appended to the value.