Don't try to xmlStrdup(NULL)
If no keychain name was given to 'cnew', it prompts for one. But before
this, we check for this condition by strtok()ing 'line', which will
return NULL after removing the command name (cnew), because there are no
more words that can be tokenized. malloc_check()ing 'name' after this
will result in a seemingly out of memory condition, because xmlStrdup()
returned NULL (for dup'ing NULL).
Only xmlStrdup() if we're sure there is a keychain name on the command
line.

Fixes #2