Enable to store use passwords, and don't rely on it being NUL terminated
Introduce a new parameter for db_parameters struct, pass_len. This holds
the db_params->pass variables length.
As a side-effect I removed the stripping of new-line chars from password
files' ends, so we'll take what's in it literally.
Now we don't need to convert the OpenSSH agent signature to HEX numbers,
we can store it as-is, and we'll know its length.

Adjust tests accordingly.