Allow spaces in the keychain names when they're a 'cnew' parameter
When specifying the keychain name directly on the command line for
'cnew', actually parse the rest of the line as the name of the keychain,
not just until the first space.

This allows creating of keychains with spaces in their names not only
buy issuing the 'cnew' command on its own then typing in the name into
the prompt, but also when adding them on the command line as a
parameter.

Basically the same change had happened a while ago for the 'new'
command. The idea here is the same.