Allow usage of a password with both SSH agent and a security key.
Now there's an option to mix in the password to the data signed by the
SSH agent -- i.e. a non-automatic usage.

Also, using a security key, one can now use it without a user-provided
challenge (i.e. a password), and use it kind of automatically with only
the salt in the database file.

'import', 'export', 'passwd' commands have been prepared as well.

Documentation is coming.