ykchalresp.c: make use of the full available length of the password
Do a little trick and given the maximum length (currently 64 bytes) of a
password when using it with a security key, always fill in the remainder
of the challenge with data from the salt after we copied the
user-supplied password there.

Also make sure *AGAIN* to always update the pass and pass_len fields in
db_param, not only when using a password with the security key...

Always try to reset and free passtmp.

A couple of whitespace fixes, while here.