Remove trailing space(s) when creating or changing keychains
Also update Changelog

There's this interesting thing with readline where it would put a space
after the completion of a word (when using TAB-complete). While the
TAB-completion implementation for editline is a custom one where I
actually replicated this and could remove it, readline is a different
beast and it doesn't seem trivial to make it not put spaces after
completed words.
So while we allow spaces in keychain names, we remove the trailing ones
as kc would not be able to change to a keychain when typing:
'c someth<TAB>'
... as it would be auto-completed to 'something ' and would
actually try to change to that (and we wouldn't know if the last space
was there intentionally in the first place).