Factor out the cmdline arg parsing into a separate function
I've always hated this code duplication and maintaining it was a
nightmare. Let's hope now that I have only one place to mess up the
parameter parsing it'll be less frequent.

export/import/passwd commands shared this code with main() in kc.c.

The help and version args work a bit funny, but oh well...