Tune multiple security key handling
If any one of the keys has the ',password' parameter, it turns on the
usage of the user-password for the security key challenge.
We'll use it only for the first key, and any subsequent keys will reuse
the previous response from the (previous) security key.
This latter goes for the non-interactive usage as well, when there's no
user-supplied password in that the first security will have the salt as
the challenge (as usual) and any subsequent key will reuse the previous
response from the (previous) security key.

We don't need a key-specific yk_pass variable for this, we're using a
"global" (i.e. database-level) one now.