Allow specifying a serial number instead of device index
Instead of a device index for the '-Y' parameter, one can specify a
serial number, and we'll try to find the corresponding device index when
we are working with security keys.
Old format still works, and if there's any garbage in the device index
field for '-Y', we're falling back to index #0.