Accommodate OpenSSL v3
We now have a return value for BIO_set_cipher() (OpenSSL v1.1.1 has
actually already had this, but the documentation was not updated
accordingly - it's all sorted in v3), so we're checking it.

To reduce code duplication and spagettification, use a function pointer
when selecting the appropriate cipher function based on the cipher name
and cipher mode.

Need for OpenSSL v3, is loading of so called providers. The 'default'
provider is loaded automatically when using some of the functions we're
using, but to make the blowfish cipher available, we need to load the
'legacy' provider. However, when loading a provider explicitly, the
'default' one doesn't load automatically. Hence we're loading both
'legacy' and 'default'.

OPENSSL_VERSION_{MAJOR, MINOR, PATCH} only exists since OpenSSL v3,
previous versions had different, more cumbersome versionings, like hex
numbers and text strings describing its version. By checking for these
macros explicitly, we're actually already checking if we're using
OpenSSL v3 or earlier.