Implement key length parameter (-K)
This allows changing the encryption key length with the '-K' parameter
for the CLI executable and for 'export', 'import' and 'passwd'
commmands.

There are two new macros, MIN_KEY_LEN and MAX_KEY_LEN which is used to
control the range of the key length. Previously it was a hard-coded 128
bytes (KEY_LEN). As none of the KDFs/ciphers used that many (1024) bits,
we can safely reduce this default to the lowest useful value (32 bytes =
256 bits) while still being able to open databases without having to
specify this new parameter.

Because there are invalid combinations of KDF+cipher+key_length, one can
lock themeselves out of their databases when used badly. Maybe we should
emit some warnings in these cases? I don't know if it's possible to
check for all invalid combinations... For now, there's just a warning
about this in the manual.

While here, bring the documentation (both in-app and the manual)
up-to-speed. There were missing parameters and badly specified
arguments (especially with '-Y').